feature : create home screen
I decided not to use Paging3 for pagination because it is too dependent on the Android platform. It also doesn’t fit well with features like adding and removing items for a future Bookmark function, and overall it doesn’t align well with Clean Architecture. Instead, I applied a custom pagination solution.